+++ ichnaea/models/cell.py
@@ -3,7 +3,6 @@ import math
 import struct
 
 import colander
-from enum import IntEnum
 from sqlalchemy import (
     BINARY,
     Column,
@@ -25,6 +24,7 @@ from ichnaea.models.base import (
     CreationMixin,
 )
 from ichnaea.models import constants
+from ichnaea.models.constants import Radio
 from ichnaea.models.sa_types import (
     TinyIntEnum,
 )
@@ -61,17 +61,6 @@ integer for the cid part.
 """
 
 CELL_SHARDS = {}
-
-
-class Radio(IntEnum):
-    """An integer enum representing a radio type."""
-    __order__ = 'gsm cdma wcdma umts lte'
-
-    gsm = 0  #: 0
-    cdma = 1  #: 1
-    wcdma = 2  #: 2
-    umts = 2  #: 2
-    lte = 3  #: 3
 
 
 def decode_cellarea(value, codec=None):
@@ -268,26 +257,75 @@ class ValidCellSignalSchema(colander.MappingSchema, ValidatorNode):
 
     asu = DefaultNode(
         colander.Integer(),
-        missing=None, validator=colander.Range(0, 97))
+        missing=None, validator=colander.Range(
+            min(constants.MIN_CELL_ASU.values()),
+            max(constants.MAX_CELL_ASU.values())))
+
     signal = DefaultNode(
         colander.Integer(),
         missing=None, validator=colander.Range(
-            constants.MIN_CELL_SIGNAL, constants.MAX_CELL_SIGNAL))
+            min(constants.MIN_CELL_SIGNAL.values()),
+            max(constants.MAX_CELL_SIGNAL.values())))
+
     ta = DefaultNode(
         colander.Integer(),
-        missing=None, validator=colander.Range(0, 63))
+        missing=None, validator=colander.Range(
+            constants.MIN_CELL_TA, constants.MAX_CELL_TA))
+
+    def _signal_from_asu(self, radio, value):
+        if radio is Radio.gsm:
+            return (value * 2) - 113
+        if radio is Radio.wcdma:
+            return value - 116
+        if radio is Radio.lte:
+            return value - 140
 
     def deserialize(self, data):
         if data:
             # Sometimes the asu and signal fields are swapped
             if (data.get('asu') is not None and
-                    data.get('asu', 0) < -1 and
-                    data.get('signal', None) == 0):
+                    data.get('asu', 0) < -5 and
+                    (data.get('signal') is None or
+                     data.get('signal', 0) >= 0)):
                 # shallow copy
                 data = dict(data)
                 data['signal'] = data['asu']
                 data['asu'] = None
-        return super(ValidCellSignalSchema, self).deserialize(data)
+
+        data = super(ValidCellSignalSchema, self).deserialize(data)
+
+        if isinstance(data.get('radio'), Radio):
+            radio = data['radio']
+
+            # Radio type specific checks for ASU field
+            if data.get('asu') is not None:
+                if not (constants.MIN_CELL_ASU[radio] <=
+                        data['asu'] <=
+                        constants.MAX_CELL_ASU[radio]):
+                    data = dict(data)
+                    data['asu'] = None
+
+            # Radio type specific checks for signal field
+            if data.get('signal') is not None:
+                if not (constants.MIN_CELL_SIGNAL[radio] <=
+                        data['signal'] <=
+                        constants.MAX_CELL_SIGNAL[radio]):
+                    data = dict(data)
+                    data['signal'] = None
+
+            # Radio type specific checks for TA field
+            if data.get('ta') is not None and radio is Radio.wcdma:
+                data = dict(data)
+                data['ta'] = None
+
+            # Calculate signal from ASU field
+            if data.get('asu') is not None and data.get('signal') is None:
+                if (constants.MIN_CELL_ASU[radio] <= data['asu'] <=
+                        constants.MAX_CELL_ASU[radio]):
+                    data = dict(data)
+                    data['signal'] = self._signal_from_asu(radio, data['asu'])
+
+        return data
 
 
 class ValidCellAreaSchema(ValidCellAreaKeySchema,