+++ ichnaea/models/tests/test_schema.py
@@ -74,14 +74,11 @@ class TestCell(ValidationTest):
             'report_id': None,
         }
         cell = {
-            'asu': 15,
             'cid': 34567,
             'lac': 12345,
             'mcc': FRANCE_MCC,
             'mnc': VIVENDI_MNC,
             'psc': None,
-            'signal': -83,
-            'ta': 5,
         }
         for (k, v) in kwargs.items():
             if k in obs:
@@ -175,58 +172,34 @@ class TestCell(ValidationTest):
                 self.check_normalized_cell(obs, cell, None)
 
     def test_invalid_latitude(self):
-        invalid_latitudes = [constants.MIN_LAT - 0.1, constants.MAX_LAT + 0.1]
-        for lat in invalid_latitudes:
+        for lat in [constants.MIN_LAT - 0.1, constants.MAX_LAT + 0.1]:
             obs, cell = self.get_sample(lat=lat)
             self.check_normalized_cell(obs, cell, None)
 
     def test_invalid_longitude(self):
-        invalid_longitudes = [constants.MIN_LON - 0.1, constants.MAX_LON + 0.1]
-        for lon in invalid_longitudes:
+        for lon in [constants.MIN_LON - 0.1, constants.MAX_LON + 0.1]:
             obs, cell = self.get_sample(lon=lon)
             self.check_normalized_cell(obs, cell, None)
 
     def test_valid_accuracy(self):
-        valid_accuracies = [0.0, 1.6, 10.1, constants.MAX_ACCURACY_CELL]
-        for accuracy in valid_accuracies:
+        for accuracy in [0.0, 1.6, 10.1, constants.MAX_ACCURACY_CELL]:
             obs, cell = self.get_sample(accuracy=accuracy)
             self.check_normalized_cell(obs, cell, {'accuracy': accuracy})
 
     def test_valid_altitude(self):
-        valid_altitudes = [-100.0, -1.6, 0.0, 10.1, 100.0]
-        for altitude in valid_altitudes:
+        for altitude in [-100.0, -1.6, 0.0, 10.1, 100.0]:
             obs, cell = self.get_sample(altitude=altitude)
             self.check_normalized_cell(obs, cell, {'altitude': altitude})
 
     def test_valid_altitude_accuracy(self):
-        valid_altitude_accuracies = [0.0, 1.6, 100.1, 1000.0]
-        for altitude_accuracy in valid_altitude_accuracies:
+        for altitude_accuracy in [0.0, 1.6, 100.1, 1000.0]:
             obs, cell = self.get_sample(
                 altitude_accuracy=altitude_accuracy)
             self.check_normalized_cell(
                 obs, cell, {'altitude_accuracy': altitude_accuracy})
 
-    def test_valid_asu(self):
-        valid_asus = [0, 10, 31, 97]
-        for asu in valid_asus:
-            obs, cell = self.get_sample(asu=asu)
-            self.check_normalized_cell(obs, cell, {'asu': asu})
-
-    def test_valid_ta(self):
-        valid_tas = [0, 15, 63]
-        for ta in valid_tas:
-            obs, cell = self.get_sample(ta=ta)
-            self.check_normalized_cell(obs, cell, {'ta': ta})
-
-    def test_valid_signal(self):
-        valid_signals = [-150, -100, -1]
-        for signal in valid_signals:
-            obs, cell = self.get_sample(signal=signal)
-            self.check_normalized_cell(obs, cell, {'signal': signal})
-
     def test_invalid_accuracy(self):
-        invalid_accuracies = [-10.0, -1.2]
-        for accuracy in invalid_accuracies:
+        for accuracy in [-10.0, -1.2]:
             obs, cell = self.get_sample(accuracy=accuracy)
             self.check_normalized_cell(obs, cell, {'accuracy': None})
 
@@ -234,40 +207,107 @@ class TestCell(ValidationTest):
         self.check_normalized_cell(obs, cell, None)
 
     def test_invalid_altitude(self):
-        invalid_altitudes = [-20000.0, 200000.0]
-        for altitude in invalid_altitudes:
+        for altitude in [-20000.0, 200000.0]:
             obs, cell = self.get_sample(altitude=altitude)
             self.check_normalized_cell(obs, cell, {'altitude': None})
 
     def test_invalid_altitude_accuracy(self):
-        invalid_altitude_accuracies = [-10.0, -1.2, 500000.0]
-        for altitude_accuracy in invalid_altitude_accuracies:
+        for altitude_accuracy in [-10.0, -1.2, 500000.0]:
             obs, cell = self.get_sample(
                 altitude_accuracy=altitude_accuracy)
             self.check_normalized_cell(
                 obs, cell, {'altitude_accuracy': None})
 
-    def test_invalid_asu(self):
-        invalid_asus = [-10, -1, 99]
-        for asu in invalid_asus:
-            obs, cell = self.get_sample(asu=asu)
-            self.check_normalized_cell(obs, cell, {'asu': None})
+    def test_valid_ta(self):
+        for ta in (constants.MIN_CELL_TA, 15, constants.MAX_CELL_TA):
+            for radio in (Radio.gsm, Radio.lte):
+                obs, cell = self.get_sample(radio=radio.name, ta=ta)
+                self.check_normalized_cell(obs, cell, {'ta': ta})
 
     def test_invalid_ta(self):
-        invalid_tas = [-10, -1, 64, 100]
-        for ta in invalid_tas:
-            obs, cell = self.get_sample(ta=ta)
+        for ta in (constants.MIN_CELL_TA - 1, constants.MAX_CELL_TA + 1):
+            for radio in (Radio.gsm, Radio.wcdma, Radio.lte):
+                obs, cell = self.get_sample(radio=radio.name, ta=ta)
+                self.check_normalized_cell(obs, cell, {'ta': None})
+
+        for ta in (constants.MIN_CELL_TA, 15, constants.MAX_CELL_TA):
+            obs, cell = self.get_sample(radio=Radio.wcdma.name, ta=ta)
             self.check_normalized_cell(obs, cell, {'ta': None})
 
+    def test_valid_asu(self):
+        for asu in (0, 10, 31):
+            obs, cell = self.get_sample(radio=Radio.gsm.name, asu=asu)
+            self.check_normalized_cell(obs, cell, {'asu': asu})
+
+        for asu in (-5, 0, 10, 91):
+            obs, cell = self.get_sample(radio=Radio.wcdma.name, asu=asu)
+            self.check_normalized_cell(obs, cell, {'asu': asu})
+
+        for asu in (0, 10, 97):
+            obs, cell = self.get_sample(radio=Radio.lte.name, asu=asu)
+            self.check_normalized_cell(obs, cell, {'asu': asu})
+
+    def test_invalid_asu(self):
+        for asu in (-5, -1, 32, 255):
+            obs, cell = self.get_sample(radio=Radio.gsm.name, asu=asu)
+            self.check_normalized_cell(obs, cell, {'asu': None})
+
+        for asu in (-10, -6, 92, 99, 255):
+            obs, cell = self.get_sample(radio=Radio.wcdma.name, asu=asu)
+            self.check_normalized_cell(obs, cell, {'asu': None})
+
+        for asu in (-5, -1, 98, 99, 255):
+            obs, cell = self.get_sample(radio=Radio.lte.name, asu=asu)
+            self.check_normalized_cell(obs, cell, {'asu': None})
+
+    def test_valid_signal(self):
+        for signal in (-113, -100, -51):
+            obs, cell = self.get_sample(radio=Radio.gsm.name, signal=signal)
+            self.check_normalized_cell(obs, cell, {'signal': signal})
+
+        for signal in (-121, -100, -25):
+            obs, cell = self.get_sample(radio=Radio.wcdma.name, signal=signal)
+            self.check_normalized_cell(obs, cell, {'signal': signal})
+
+        for signal in (-140, -100, -43):
+            obs, cell = self.get_sample(radio=Radio.lte.name, signal=signal)
+            self.check_normalized_cell(obs, cell, {'signal': signal})
+
     def test_invalid_signal(self):
-        invalid_signals = [-300, -151, 0, 10]
-        for signal in invalid_signals:
-            obs, cell = self.get_sample(signal=signal)
+        for signal in (-114, -50, 0, 10):
+            obs, cell = self.get_sample(radio=Radio.gsm.name, signal=signal)
+            self.check_normalized_cell(obs, cell, {'signal': None})
+
+        for signal in (-122, -24, 0, 10):
+            obs, cell = self.get_sample(radio=Radio.wcdma.name, signal=signal)
+            self.check_normalized_cell(obs, cell, {'signal': None})
+
+        for signal in (-141, -42, 0, 10):
+            obs, cell = self.get_sample(radio=Radio.lte.name, signal=signal)
             self.check_normalized_cell(obs, cell, {'signal': None})
 
     def test_asu_signal_field_mix(self):
         obs, cell = self.get_sample(asu=-75, signal=0)
         self.check_normalized_cell(obs, cell, {'signal': -75})
+
+    def test_asu_signal_conversion(self):
+        for asu, signal in ((-1, None), (0, -113), (16, -81),
+                            (31, -51), (32, None)):
+            obs, cell = self.get_sample(
+                radio=Radio.gsm.name, asu=asu, signal=None)
+            self.check_normalized_cell(obs, cell, {'signal': signal})
+
+        for asu, signal in ((-6, None), (-5, -121), (0, -116),
+                            (16, -100), (91, -25), (92, None)):
+            obs, cell = self.get_sample(
+                radio=Radio.wcdma.name, asu=asu, signal=None)
+            self.check_normalized_cell(obs, cell, {'signal': signal})
+
+        for asu, signal in ((-1, None), (0, -140), (40, -100),
+                            (97, -43), (98, None)):
+            obs, cell = self.get_sample(
+                radio=Radio.lte.name, asu=asu, signal=None)
+            self.check_normalized_cell(obs, cell, {'signal': signal})
 
     def test_cid_65535_invalid_lac(self):
         obs, cell = self.get_sample(lac=None, cid=65535, psc=1)
@@ -283,7 +323,6 @@ class TestCell(ValidationTest):
             {'lac': 3, 'cid': None},
             {'lac': None, 'cid': 4},
         ]
-
         for entry in entries:
             obs, cell = self.get_sample(**entry)
             self.check_normalized_cell(obs, cell, None)
@@ -391,14 +430,12 @@ class TestWifi(ValidationTest):
             self.check_normalized_wifi(obs, wifi, None)
 
     def test_valid_accuracy(self):
-        valid_accuracies = [0.0, 1.6, 10.1, constants.MAX_ACCURACY_WIFI]
-        for accuracy in valid_accuracies:
+        for accuracy in [0.0, 1.6, 10.1, constants.MAX_ACCURACY_WIFI]:
             obs, wifi = self.get_sample(accuracy=accuracy)
             self.check_normalized_wifi(obs, wifi, {'accuracy': accuracy})
 
     def test_invalid_accuracy(self):
-        invalid_accuracies = [-10.0, -1.2]
-        for accuracy in invalid_accuracies:
+        for accuracy in [-10.0, -1.2]:
             obs, wifi = self.get_sample(accuracy=accuracy)
             self.check_normalized_wifi(obs, wifi, {'accuracy': None})
 
@@ -410,6 +447,7 @@ class TestWifi(ValidationTest):
             (2412, 1),
             (2427, 4),
             (2472, 13),
+            (2484, 14),
             (5170, 34),
             (5200, 40),
             (5805, 161),
@@ -433,46 +471,35 @@ class TestWifi(ValidationTest):
             self.assertFalse('frequency' in wifi)
 
     def test_valid_signal(self):
-        valid_signals = [-100, -99, -1]
-
-        for signal in valid_signals:
+        for signal in (constants.MIN_WIFI_SIGNAL, constants.MAX_WIFI_SIGNAL):
             obs, wifi = self.get_sample(signal=signal)
             self.check_normalized_wifi(obs, wifi, dict(signal=signal))
 
     def test_invalid_signal(self):
-        invalid_signals = [-300, -101, 0, 10]
-
-        for signal in invalid_signals:
+        for signal in (constants.MIN_WIFI_SIGNAL - 1,
+                       constants.MAX_WIFI_SIGNAL + 1):
             obs, wifi = self.get_sample(signal=signal)
             self.check_normalized_wifi(obs, wifi, dict(signal=None))
 
     def test_valid_snr(self):
-        valid_snrs = [0, 12, 100]
-
-        for snr in valid_snrs:
+        for snr in [0, 12, 100]:
             obs, wifi = self.get_sample(snr=snr)
             self.check_normalized_wifi(obs, wifi, dict(snr=snr))
 
     def test_invalid_snr(self):
-        invalid_snrs = [-1, -50, 101]
-
-        for snr in invalid_snrs:
+        for snr in [-1, -50, 101]:
             obs, wifi = self.get_sample(snr=snr)
             self.check_normalized_wifi(obs, wifi, dict(snr=None))
 
     def test_valid_channel(self):
-        valid_channels = [1, 20, 45, 165]
-
-        for channel in valid_channels:
+        for channel in [1, 20, 45, 165]:
             obs, wifi = self.get_sample(channel=channel)
             wifi = self.check_normalized_wifi(
                 obs, wifi, dict(channel=channel))
             self.assertFalse('frequency' in wifi)
 
     def test_invalid_channel_valid_frequency(self):
-        invalid_channels = [-10, -1, 201, 2500]
-
-        for channel in invalid_channels:
+        for channel in [-10, -1, 201, 2500]:
             obs, wifi = self.get_sample()
             chan = wifi['channel']
             wifi['channel'] = channel
@@ -481,9 +508,7 @@ class TestWifi(ValidationTest):
             self.assertFalse('frequency' in wifi)
 
     def test_invalid_frequency(self):
-        invalid_frequencies = [-1, 2000, 2411, 2473, 5168, 5826, 6000]
-
-        for frequency in invalid_frequencies:
+        for frequency in [-1, 2000, 2411, 2473, 5168, 5826, 6000]:
             obs, wifi = self.get_sample(frequency=frequency)
             chan = wifi['channel']
             wifi = self.check_normalized_wifi(obs, wifi,